refactor: rename 'utils' package to 'components'
