+++ app/src/main/kotlin/view/components/RadioColumn.kt
@@ -1,4 +1,4 @@
-package view.utils
+package view.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
