+++ app/src/main/kotlin/view/components/ToolBox.kt
@@ -1,4 +1,4 @@
-package view.utils
+package view.components
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
