+++ app/src/main/kotlin/view/components/dialogWindows/CreateGraphDialogWindow.kt
@@ -1,21 +1,19 @@
-package view.utils.dialogWindows
+package view.components.dialogWindows
 
 import MyAppTheme
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
-import view.utils.RadioColumn
+import view.components.RadioColumn
 import viewmodel.graph.SetupGraphViewModel
 import viewmodel.graph.createGraphFromTypesIndices
