+++ app/src/main/kotlin/view/components/dialogWindows/ErrorWindow.kt
@@ -1,4 +1,4 @@
-package view.utils.dialogWindows
+package view.components.dialogWindows
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.Button
@@ -6,8 +6,6 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
