+++ app/src/main/kotlin/view/components/dialogWindows/ImportGraphDialogWindow.kt
@@ -1,4 +1,4 @@
-package view.utils.dialogWindows
+package view.components.dialogWindows
 
 import JSON
 import MyAppTheme
