+++ app/src/main/kotlin/view/components/dialogWindows/Neo4jImportGraphDialogWindow.kt
@@ -1,4 +1,4 @@
-package view.utils.dialogWindows
+package view.components.dialogWindows
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
