+++ app/src/main/kotlin/view/components/dialogWindows/SQLiteImportGraphDialogWindow.kt
@@ -1,4 +1,4 @@
-package view.utils.dialogWindows
+package view.components.dialogWindows
 
 import model.io.sql.SQLDatabaseModule
 import androidx.compose.foundation.background
