+++ app/src/main/kotlin/view/tabScreen/FileControlTab.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.window.Dialog
 import kotlinx.coroutines.delay
 import model.io.neo4j.Neo4jRepositoryHandler
 import model.io.sql.SQLDatabaseModule
-import view.utils.dialogWindows.*
+import view.components.dialogWindows.*
 import viewmodel.graph.GraphViewModel
 import java.awt.FileDialog
 import java.awt.Frame
