+++ app/src/main/kotlin/view/tabScreen/analyzeTab/algorithmsUI/CyclesUI.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.unit.sp
 import view.tabScreen.analyzeTab.borderPadding
 import view.tabScreen.analyzeTab.horizontalGap
 import view.tabScreen.analyzeTab.rowHeight
-import view.utils.dialogWindows.ErrorWindow
+import view.components.dialogWindows.ErrorWindow
 import viewmodel.graph.GraphViewModel
 
 @Composable
